+++ src/test/java/calculator/number/Number.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.number;
 
 public interface Number {
