+++ src/test/java/calculator/number/PositiveNumber.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.number;
 
 public class PositiveNumber implements Number {
